gateway: make buffer request mem and max configurable
This patch provides proper implementation for the previous patch. It
allows to configure the request buffer size and max body size instead of
using hard coded values.

The default are now:

- buffer size: 1MB
- max body size: 100MB

Since it is not possible to have limits based on routes, this represents
the absolute safe limit the gateway can proxy, and 1MB was too low.